Adjustments for PR
