+++ docker-compose.db.yaml
@@ -17,8 +17,8 @@ services:
     ports:
       - 5050:80
     environment:
-      PGADMIN_DEFAULT_EMAIL: ${PGADMIN_DEFAULT_EMAIL}
-      PGADMIN_DEFAULT_PASSWORD: ${PGADMIN_DEFAULT_PASSWORD}
+      PGADMIN_DEFAULT_EMAIL: "admin@admin.com"
+      PGADMIN_DEFAULT_PASSWORD: "1234"
     restart: "no"
 volumes:
   db_data:
