+++ frontend/src/aspects/Auth/EmailVerificationRequiredPage.tsx
@@ -36,11 +36,10 @@ export default function EmailVerificationRequiredPage({
             {!message && !alreadyUsed ? (
                 <Text fontSize="xl" lineHeight="revert" fontWeight="light" maxW={600}>
                     <FormattedMessage
-                        id="auth.emailverificationrequiredpage.pleaseverify"
+                        id="auth.emailverificationrequiredpage.emailinstructions"
                         defaultMessage="
-                            Before you can login you must verify your email address<br>
-                            <b>You should have received an email from our domain</b>
-                            &nbsp;with your verification link. After verifying your email, please log in to continue.
+                            Before you can login you must verify your email address. 
+                            You should have received an email from our domain with your verification link. After verifying your email, please log in to continue.
                         "
                     />
                 </Text>
