+++ frontend/src/index.tsx
@@ -14,34 +14,63 @@ import ApolloCustomProvider from "./aspects/GQL/ApolloCustomProvider";
 import { UXChoiceProvider } from "./aspects/UXChoice/UXChoice";
 import UXChoiceDialog from "./aspects/UXChoice/UXChoiceDialog";
 import "./index.css";
+import { IntlProvider } from 'react-intl';
+import messages_pt from "./lang-compiled/pt.json";
+import messages_en from "./lang-compiled/en.json";
+import messages_de from "./lang-compiled/de.json";
+import messages_it from "./lang-compiled/it.json";
+
+function loadLocaleData(locale: string) {
+    switch (locale) {
+        case 'pt':
+        case 'pt-br':
+        case 'pt-BR':
+            return messages_pt;
+        case 'de':
+        case 'de-DE':
+            return messages_de;
+        case 'it':
+        case 'it-IT':
+            return messages_it;
+        default:
+            return messages_en;
+    }
+}
 
 async function bootstrapApplication() {
-    const [locale, setLocale] = useRestorableState<string>(
-        "clowdr-language",
-        'en',
-        (x) => JSON.stringify(x),
-        (x) => JSON.parse(x)
-    );
+    // const [locale, setLocale] = useRestorableState<string>(
+    //     "clowdr-language",
+    //     'en',
+    //     (x) => JSON.stringify(x),
+    //     (x) => JSON.parse(x)
+    // );
+    const locale = navigator.language || 'en';
     ReactDOM.render(
         <React.StrictMode>
-            <ErrorBoundary FallbackComponent={AppError}>
-                    <VonageGlobalStateProvider>
-                        <HelmetProvider>
-                            <BrowserRouter>
-                                <Auth0CustomProvider>
-                                    <ApolloCustomProvider>
-                                        <ChakraCustomProvider>
-                                            <UXChoiceProvider>
-                                                <UXChoiceDialog />
-                                                <App locale={locale} />
-                                            </UXChoiceProvider>
-                                        </ChakraCustomProvider>
-                                    </ApolloCustomProvider>
-                                </Auth0CustomProvider>
-                            </BrowserRouter>
-                        </HelmetProvider>
-                    </VonageGlobalStateProvider>
-            </ErrorBoundary>
+            <IntlProvider
+                locale={locale}
+                defaultLocale="en"
+                messages={loadLocaleData(locale)}
+            >
+                <ErrorBoundary FallbackComponent={AppError}>
+                        <VonageGlobalStateProvider>
+                            <HelmetProvider>
+                                <BrowserRouter>
+                                    <Auth0CustomProvider>
+                                        <ApolloCustomProvider>
+                                            <ChakraCustomProvider>
+                                                <UXChoiceProvider>
+                                                    <UXChoiceDialog />
+                                                    <App locale={locale} />
+                                                </UXChoiceProvider>
+                                            </ChakraCustomProvider>
+                                        </ApolloCustomProvider>
+                                    </Auth0CustomProvider>
+                                </BrowserRouter>
+                            </HelmetProvider>
+                        </VonageGlobalStateProvider>
+                </ErrorBoundary>
+            </IntlProvider>
         </React.StrictMode>,
         document.getElementById("root"));
 }
