+++ frontend/src/lang-compiled/de.json
@@ -1,4 +1,240 @@
 {
+  "auth.emailverificationrequiredpage.emailinstructions": [
+  ],
+  "auth.emailverificationrequiredpage.emailverified": [
+    {
+      "type": 0,
+      "value": "Ihre E-Mail wurde bestätigt. Sie können sich jetzt einloggen."
+    }
+  ],
+  "auth.emailverificationrequiredpage.pleaseverify": [
+    {
+      "type": 0,
+      "value": "Bitte bestätige deine Email Adresse"
+    }
+  ],
+  "auth.emailverificationrequiredpage.title": [
+    {
+      "type": 0,
+      "value": "E-Mail-Bestätigung erforderlich"
+    }
+  ],
+  "auth.emailverificationrequiredpage.youcanclose": [
+    {
+      "type": 0,
+      "value": "Sie können diese Seite jetzt schließen, während Sie auf Ihre Bestätigungs-E-Mail warten."
+    }
+  ],
+  "auth.loggedoutpage.loadinguserinfo": [
+    {
+      "type": 0,
+      "value": "Benutzerinformationen werden geladen"
+    }
+  ],
+  "auth.loggedoutpage.loggedout": [
+    {
+      "type": 0,
+      "value": "Sie wurden ausgeloggt."
+    }
+  ],
+  "auth.loggedoutpage.title": [
+    {
+      "type": 0,
+      "value": "Abgemeldet"
+    }
+  ],
+  "auth.passwordresetresultpage.passwordreset": [
+    {
+      "type": 0,
+      "value": "Ihr Passwort wurde zurückgesetzt"
+    }
+  ],
+  "auth.passwordresetresultpage.statusunknown": [
+    {
+      "type": 0,
+      "value": "Passwort zurücksetzen - Status unbekannt"
+    }
+  ],
+  "auth.protectedroute.logout": [
+    {
+      "type": 0,
+      "value": "Ausloggen"
+    }
+  ],
+  "auth.protectedroute.usernotfound": [
+    {
+      "type": 0,
+      "value": "Benutzerkonto wurde nicht gefunden. Bitte wenden Sie sich an den technischen Support, wenn dieses Problem weiterhin besteht."
+    }
+  ],
+  "badges.badgeinput.badgenotfound": [
+    {
+      "type": 0,
+      "value": "Fehler! Abzeichen nicht gefunden"
+    }
+  ],
+  "badges.badgeinput.badges": [
+    {
+      "type": 0,
+      "value": "Abzeichen"
+    }
+  ],
+  "badges.badgeinput.createbadges": [
+    {
+      "type": 0,
+      "value": "Erstellen Sie Abzeichen durch Tippen, drücken Sie die Eingabetaste und wählen Sie dann eine Farbe aus."
+    }
+  ],
+  "badges.badgeinput.delete": [
+    {
+      "type": 0,
+      "value": "Löschen Sie Abzeichen, indem Sie darauf klicken."
+    }
+  ],
+  "badges.badgeinput.done": [
+    {
+      "type": 0,
+      "value": "Erledigt"
+    }
+  ],
+  "badges.badgeinput.example01": [
+    {
+      "type": 0,
+      "value": "Autor"
+    }
+  ],
+  "badges.badgeinput.example02": [
+    {
+      "type": 0,
+      "value": "Freiwillige"
+    }
+  ],
+  "badges.badgeinput.example03": [
+    {
+      "type": 0,
+      "value": "Mentor"
+    }
+  ],
+  "badges.badgeinput.example04": [
+    {
+      "type": 0,
+      "value": "Mentee"
+    }
+  ],
+  "badges.badgeinput.example05": [
+    {
+      "type": 0,
+      "value": "Sponsor"
+    }
+  ],
+  "badges.badgeinput.example06": [
+    {
+      "type": 0,
+      "value": "Einstellung"
+    }
+  ],
+  "badges.badgeinput.example07": [
+    {
+      "type": 0,
+      "value": "Stellt mich ein"
+    }
+  ],
+  "badges.badgeinput.example08": [
+    {
+      "type": 0,
+      "value": "Lass uns schreiben!"
+    }
+  ],
+  "badges.badgeinput.example09": [
+    {
+      "type": 0,
+      "value": "Kaffee?"
+    }
+  ],
+  "badges.badgeinput.example10": [
+    {
+      "type": 0,
+      "value": "Veranstalter"
+    }
+  ],
+  "badges.badgeinput.examples": [
+    {
+      "type": 0,
+      "value": "Beispiele (zum Hinzufügen anklicken)"
+    }
+  ],
+  "badges.badgeinput.maxlength": [
+    {
+      "type": 0,
+      "value": "Maximale Länge 10 Zeichen."
+    }
+  ],
+  "badges.badgeinput.selectcolour": [
+    {
+      "type": 0,
+      "value": "Farbe des Abzeichens auswählen"
+    }
+  ],
+  "badges.badgeinput.selectedbadges": [
+    {
+      "type": 0,
+      "value": "Ausgewählte Abzeichen"
+    }
+  ],
+  "chat.compose.chatcompose.invalidfilename": [
+    {
+      "type": 0,
+      "value": "Ungültiger Dateiname. Dateinamen dürfen nur Buchstaben, Zahlen, Leerzeichen und folgende Sonderzeichen enthalten: !*'()-_"
+    }
+  ],
+  "chat.compose.chatcompose.sendfailedmessage": [
+    {
+      "type": 0,
+      "value": "Ihre Nachricht konnte nicht gesendet werden. Bitte versuchen Sie es gleich noch einmal. Wenn der unten angegebene Fehler weiterhin besteht, wenden Sie sich bitte an unseren technischen Support."
+    }
+  ],
+  "chat.compose.chatcompose.sendfailedtitle": [
+    {
+      "type": 0,
+      "value": "Senden fehlgeschlagen"
+    }
+  ],
+  "chat.compose.chatcompose.uploadfailed": [
+    {
+      "type": 0,
+      "value": "Fehler beim Hochladen der Datei. Bitte versuchen Sie es später erneut."
+    }
+  ],
+  "chat.compose.poll.createpolloptionsmodal.options": [
+    {
+      "type": 0,
+      "value": "Optionen"
+    }
+  ],
+  "chat.compose.poll.createpolloptionsmodal.optionshelptext": [
+    {
+      "type": 0,
+      "value": "Optionen durch Eingabe hinzufügen oder löschen. Drücken Sie die Eingabetaste, wenn Sie eine Auswahl bearbeiten, um eine neue Option darunter zu erstellen. Sie können Registranten auch ermöglichen, ihre eigenen Optionen zu erstellen."
+    }
+  ],
+  "chat.compose.poll.createpolloptionsmodal.yourquestion": [
+    {
+      "type": 0,
+      "value": "Ihre Frage:"
+    }
+  ],
+  "menu.v2.collapse": [
+    {
+      "type": 0,
+      "value": "Menü reduzieren"
+    }
+  ],
+  "menu.v2.expand": [
+    {
+      "type": 0,
+      "value": "Menü erweitern"
+    }
+  ],
   "signup": [
     {
       "type": 0,
@@ -27,6 +263,54 @@
     {
       "type": 0,
       "value": "Willkommen bei Midspace"
+    }
+  ],
+  "uxchoice.uxchoicedialog.buttonmsg": [
+    {
+      "type": 0,
+      "value": "Sie können Ihre Meinung jederzeit ändern, indem Sie die Schaltfläche \"Benutzeroberfläche wechseln\" im Profilmenü verwenden."
+    }
+  ],
+  "uxchoice.uxchoicedialog.changedyourmind": [
+    {
+      "type": 0,
+      "value": "Ihre Meinung geändert? Kein Problem!"
+    }
+  ],
+  "uxchoice.uxchoicedialog.newxp": [
+    {
+      "type": 0,
+      "value": "Nutze die neue Erfahrung"
+    }
+  ],
+  "uxchoice.uxchoicedialog.oldxp": [
+    {
+      "type": 0,
+      "value": "Bleib beim Alten"
+    }
+  ],
+  "uxchoice.uxchoicedialog.quickchoice": [
+    {
+      "type": 0,
+      "value": "Hallo, wir haben eine schnelle Auswahl für Sie…"
+    }
+  ],
+  "uxchoice.uxchoicedialog.welcome1": [
+    {
+      "type": 0,
+      "value": "Neu bei Midspace? Wir empfehlen unser neues, einfacheres Erlebnis"
+    }
+  ],
+  "uxchoice.uxchoicedialog.welcome2": [
+    {
+      "type": 0,
+      "value": "Alle die gleichen Funktionen, nur einfacher zu navigieren."
+    }
+  ],
+  "uxchoice.uxchoicedialog.welcome3": [
+    {
+      "type": 0,
+      "value": "Habe vor kurzem Midspace verwendet und würdest lieber bei der alten Version bleiben?"
     }
   ]
 }
